Replaced some static_casts with NumericCast() and ToUnderlying() 04/27/2023 | 23w17a3

+++ Games/Headless-Tests/src/Compute/ComputeTests.cpp
@@ -18,7 +18,7 @@ void ComputeTests::OnAttach()
 
     //Load Quad vertices
     m_vertexBuffer = TRAP::Graphics::VertexBuffer::Create(m_quadVerticesIndexed.data(),
-                                                          static_cast<uint32_t>(m_quadVerticesIndexed.size()) *
+                                                          m_quadVerticesIndexed.size() *
                                                           sizeof(float), TRAP::Graphics::UpdateFrequency::Static);
     const TRAP::Graphics::VertexBufferLayout layout =
     {
@@ -30,7 +30,7 @@ void ComputeTests::OnAttach()
 
     //Load Quad indices
     m_indexBuffer = TRAP::Graphics::IndexBuffer::Create(m_quadIndices.data(),
-                                                        static_cast<uint16_t>(m_quadIndices.size()) *
+                                                        m_quadIndices.size() *
                                                         sizeof(uint16_t), TRAP::Graphics::UpdateFrequency::Static);
     m_indexBuffer->AwaitLoading();
